Struct twilight_gateway_queue::LocalQueue

pub struct LocalQueue(_);

A local, in-process implementation of a Queue which manages the connection attempts of one or more shards.

The queue will take incoming requests and then queue them, releasing one of the requests every 6 seconds. The queue is necessary because there’s a ratelimit on how often shards can initiate sessions.

You usually won’t need to handle this yourself, because the Cluster will do that for you when managing multiple shards.

When not to use this

This queue implementation is “local”, meaning it’s intended to be used if you manage shards only in this process. If you run shards in multiple different processes (do you utilize microservices a lot?), then you must not use this implementation. Shards across multiple processes may create new sessions at the same time, which is bad.

It should also not be used for very large sharding, for that the LargeBotQueue can be used.

If you can’t use this, look into an alternative implementation of the Queue, such as the gateway-queue broker.

Implementations

impl LocalQueue

pub fn new() -> Self

Creates a new local queue.