Trait twilight_gateway_queue::Queue

pub trait Queue {
    // Required method
    fn enqueue(&self, id: u32) -> Receiver<()>;
}

Abstraction for types processing gateway identify requests.

Required Methods

fn enqueue(&self, id: u32) -> Receiver<()>

Enqueue a shard with this ID.

Send () to signal the shard to proceed. Note that shards may have dropped the receiver prior.

Closing the channel should causes the shard to requeue.